Resolve "VC builds fail again"

Closes #221

See merge request hpc-against-corona/epidemiology!142

+++ cpp/epidemiology/abm/person.h
@@ -58,7 +58,7 @@ public:
 
 private:
     InfectionState m_state;
-    float m_time_until_carrier;
+    double m_time_until_carrier;
     std::reference_wrapper<Location> m_location;
     //age, times, ...
 };
